Update generator serialization to have h0 = G
For compatibility with exisiting Pedersen commitments data structures,
it is necessary to have commitments be of the G_a*v + G*gamma where G_a
is asset gen and gamma is blinding factors.

However, in BP++ design, the blinding values are along H_vec. In order
make these compatible with BP++, we make h0 = G